requirements.txt
accelerate==0.12.0
diffusers==0.9.0
ftfy==6.1.1
Pillow==9.2.0
torch==1.12.1
transformers==4.22.1
